+++ app/(dashboard)/tasks/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import GlobalTaskBoard from "../../components/GlobalTaskBoard";
+import { useProjects } from "../../hooks/useProjects";
+
+export default function TasksPage() {
+  const { projects } = useProjects();
+  return <GlobalTaskBoard projects={projects} />;
+}